+++ .gitignore
@@ -18,3 +18,4 @@ src/DataStandardize/ExpConfig.py
 .jpg
 /new_dash/temp_files/
 /new_dash/pages/file_system_store/
+.h5
