+++ Analysis/Feb2021/entropy_gamma_final.py
@@ -1,23 +1,21 @@
 """Hopefully all the final analysis of Gamma Broadened Entropy measurements"""
-import src.UsefulFunctions as U
-import src.Characters as C
 from src.DatObject.Make_Dat import get_dat, get_dats, DatHDF
-from src.DatObject.Attributes.SquareEntropy import square_wave_time_array
-from src.DatObject.Attributes.Transition import i_sense, i_sense_digamma
-from src.HDF_Util import DatDataclassTemplate
 from src.Plotting.Plotly.PlotlyUtil import additional_data_dict_converter, HoverInfo
-from src.Dash.DatPlotting import OneD, TwoD
-from Analysis.Feb2021.common import get_deltaT, plot_fit_integrated_comparison, entropy_vs_time_trace, \
-    entropy_vs_time_fig, do_narrow_fits, do_entropy_calc, do_transition_only_calc, set_sf_from_transition, \
-    calculate_csq_map, setup_csq_dat, NRG_fitter, get_integrated_trace, get_integrated_fig, transition_trace, \
-    transition_fig, single_transition_trace
+from src.Dash.DatPlotting import OneD
 
-from typing import List, Optional, Callable, Union, Tuple
-import h5py
-from progressbar import progressbar
+from Analysis.Feb2021.common import plot_fit_integrated_comparison, do_entropy_calc, do_transition_only_calc, \
+    calculate_csq_map, setup_csq_dat, get_integrated_trace, get_integrated_fig, transition_trace, \
+    transition_fig
+from src.AnalysisTools.gamma_entropy import GammaAnalysisParams, save_gamma_analysis_params_to_dat
+from src.AnalysisTools.fitting import get_centers, set_centers, calculate_tonly_data, calculate_se_output, calculate_fit
+import src.UsefulFunctions as U
+from src.Plotting.Plotly.PlotlyUtil import get_slider_figure
+from src.DatObject.Attributes.SquareEntropy import entropy_signal
+from src.Characters import DELTA
+
+from typing import List, Optional, Callable, Union, Tuple, Iterable
 import logging
 import lmfit as lm
-import plotly.graph_objects as go
 import plotly.io as pio
 import numpy as np
 from functools import partial
@@ -42,68 +40,6 @@ def reset_dats(*args: Union[list, int, None]):
 
 
 @dataclass
-class GammaAnalysisParams(DatDataclassTemplate):
-    """All the various things that go into calculating Entropy etc"""
-    # To save in HDF
-    save_name: str
-
-    # For Entropy calculation (and can also determine transition info from entropy dat)
-    entropy_datnum: int
-    setpoint_start: Optional[float] = 0.005  # How much data to throw away after each heating setpoint change in secs
-    entropy_transition_func_name: str = 'i_sense'  # For centering data only if Transition specific dat supplied,
-    # Otherwise this also is used to calculate gamma, amplitude and optionally dT
-    entropy_fit_width: Optional[float] = None
-    entropy_data_rows: tuple = (None, None)  # Only fit between rows specified (None means beginning or end)
-
-    # Integrated Entropy
-    force_dt: Optional[float] = None  # 1.11  #  dT for scaling
-    # (to determine from Entropy set dt_from_square_transition = True)
-    force_amp: Optional[float] = None  # Otherwise determined from Transition Only, or from Cold part of Entropy
-    sf_from_square_transition: bool = False  # Set True to determine dT from Hot - Cold theta
-
-    # For Transition only fitting (determining amplitude and gamma)
-    transition_only_datnum: Optional[int] = None  # For determining amplitude and gamma
-    transition_func_name: str = 'i_sense_digamma'
-    transition_center_func_name: Optional[str] = None  # Which fit func to use for centering data
-    # (defaults to same as transition_center_func_name)
-    transition_fit_width: Optional[float] = None
-    force_theta: Optional[float] = None  # 3.9  # Theta must be forced in order to get an accurate gamma for broadened
-    force_gamma: Optional[float] = None  # Gamma must be forced zero to get an accurate theta for weakly coupled
-    transition_data_rows: Tuple[Optional[int], Optional[int]] = (
-        None, None)  # Only fit between rows specified (None means beginning or end)
-
-    # For CSQ mapping, applies to both Transition and Entropy (since they should always be taken in pairs anyway)
-    csq_mapped: bool = False  # Whether to use CSQ mapping
-    csq_datnum: Optional[int] = None  # CSQ trace to use for CSQ mapping
-
-    def __str__(self):
-        return f'Dat{self.entropy_datnum}:\n' \
-               f'\tEntropy Params:\n' \
-               f'entropy datnum = {self.entropy_datnum}\n' \
-               f'setpoint start = {self.setpoint_start}\n' \
-               f'transition func = {self.entropy_transition_func_name}\n' \
-               f'force theta = {self.force_theta}\n' \
-               f'force gamma = {self.force_gamma}\n' \
-               f'fit width = {self.entropy_fit_width}\n' \
-               f'data rows = {self.entropy_data_rows}\n' \
-               f'\tIntegrated Entropy Params:\n' \
-               f'from Entropy directly = {self.sf_from_square_transition}\n' \
-               f'forced dT = {self.force_dt}\n' \
-               f'forced amp = {self.force_amp}\n' \
-               f'\tTransition Params:\n' \
-               f'transition datnum = {self.transition_only_datnum}\n' \
-               f'fit func = {self.transition_func_name}\n' \
-               f'center func = {self.transition_center_func_name}\n' \
-               f'fit width = {self.transition_fit_width}\n' \
-               f'force theta = {self.force_theta}\n' \
-               f'force gamma = {self.force_gamma}\n' \
-               f'data rows = {self.transition_data_rows}\n' \
-               f'\tCSQ Mapping Params:\n' \
-               f'Mapping used: {self.csq_mapped}\n' \
-               f'csq datnum: {self.csq_datnum}\n'
-
-
-@dataclass
 class AnalysisGeneral:
     params_list: List[GammaAnalysisParams]
 
@@ -123,57 +59,42 @@ class AnalysisGeneral:
         self.csq_datnums = [params.csq_datnum for params in self.params_list if params.csq_datnum is not None]
 
 
-def save_gamma_analysis_params_to_dat(dat: DatHDF, analysis_params: GammaAnalysisParams,
-                                      name: str):
-    """Save GammaAnalysisParams to suitable place in DatHDF"""
-    with h5py.File(dat.hdf.hdf_path, 'r+') as hdf:
-        analysis_group = hdf.require_group('Gamma Analysis')
-        analysis_params.save_to_hdf(analysis_group, name=name)
-
-
-def load_gamma_analysis_params(dat: DatHDF, name: str) -> GammaAnalysisParams:
-    """Load GammaAnalysisParams from DatHDF"""
-    with h5py.File(dat.hdf.hdf_path, 'r') as hdf:
-        analysis_group = hdf.get('Gamma Analysis')
-        analysis_params = GammaAnalysisParams.from_hdf(analysis_group, name=name)
-    return analysis_params
-
-
-def process_single(params: GammaAnalysisParams, overwrite_transition=False, overwrite_entropy=False):
+def process_single(pars: GammaAnalysisParams, overwrite_transition=False, overwrite_entropy=False):
     """Does all the processing necessary for a single GammaAnalysisParams (i.e. csq mapping, transition fitting,
     entropy fitting and setting integration info"""
-    if params.csq_mapped:
-        calculate_csq_map(params.entropy_datnum, csq_datnum=params.csq_datnum, overwrite=overwrite_entropy)
-        if params.transition_only_datnum:
-            calculate_csq_map(params.transition_only_datnum, csq_datnum=params.csq_datnum,
+    if pars.csq_mapped:
+        calculate_csq_map(pars.entropy_datnum, csq_datnum=pars.csq_datnum, overwrite=overwrite_entropy)
+        if pars.transition_only_datnum:
+            calculate_csq_map(pars.transition_only_datnum, csq_datnum=pars.csq_datnum,
                               overwrite=overwrite_transition)
 
-    if params.transition_only_datnum is not None:
-        if params.save_name + '_cold' not in get_dat(params.transition_only_datnum).Transition.fit_names \
+    if pars.transition_only_datnum is not None:
+        if pars.save_name + '_cold' not in get_dat(pars.transition_only_datnum).Transition.fit_names \
                 or overwrite_transition:
-            do_transition_only_calc(datnum=params.transition_only_datnum, save_name=params.save_name,
-                                    theta=params.force_theta, gamma=params.force_gamma,
-                                    center_func=params.transition_center_func_name,
-                                    width=params.transition_fit_width, t_func_name=params.transition_func_name,
-                                    csq_mapped=params.csq_mapped, data_rows=params.transition_data_rows,
+            do_transition_only_calc(datnum=pars.transition_only_datnum, save_name=pars.save_name,
+                                    theta=pars.force_theta, gamma=pars.force_gamma,
+                                    center_func=pars.transition_center_func_name,
+                                    width=pars.transition_fit_width, t_func_name=pars.transition_func_name,
+                                    csq_mapped=pars.csq_mapped, data_rows=pars.transition_data_rows,
                                     overwrite=overwrite_transition)
-    if params.save_name not in get_dat(params.entropy_datnum).Entropy.fit_names or overwrite_entropy:
-        do_entropy_calc(params.entropy_datnum, save_name=params.save_name,
-                        setpoint_start=params.setpoint_start,
-                        t_func_name=params.entropy_transition_func_name,
-                        csq_mapped=params.csq_mapped,
-                        theta=params.force_theta, gamma=params.force_gamma, width=params.entropy_fit_width,
-                        data_rows=params.transition_data_rows,
+    if pars.save_name not in get_dat(pars.entropy_datnum).Entropy.fit_names or overwrite_entropy:
+        do_entropy_calc(pars.entropy_datnum, save_name=pars.save_name,
+                        setpoint_start=pars.setpoint_start,
+                        t_func_name=pars.entropy_transition_func_name,
+                        csq_mapped=pars.csq_mapped,
+                        theta=pars.force_theta, gamma=pars.force_gamma, width=pars.entropy_fit_width,
+                        data_rows=pars.transition_data_rows,
                         overwrite=overwrite_entropy)
 
-    calculate_new_sf_only(params.entropy_datnum, params.save_name,
-                          dt=params.force_dt, amp=params.force_amp,
-                          from_square_transition=params.sf_from_square_transition,
-                          transition_datnum=params.transition_only_datnum,
-                          fit_name=params.save_name)
+    calculate_new_sf_only(pars.entropy_datnum, pars.save_name,
+                          dt=pars.force_dt, amp=pars.force_amp,
+                          from_square_transition=pars.sf_from_square_transition,
+                          transition_datnum=pars.transition_only_datnum,
+                          fit_name=pars.save_name)
 
     # Save to HDF
-    save_gamma_analysis_params_to_dat(dat, analysis_params=params, name=params.save_name)
+    d = get_dat(pars.entropy_datnum)
+    save_gamma_analysis_params_to_dat(d, analysis_params=pars, name=pars.save_name)
 
 
 def calculate_new_sf_only(entropy_datnum: int, save_name: str,
@@ -242,7 +163,7 @@ class PlotInfo:
     trace_name: Callable
 
 
-def plot_stacked_square_heated(datnums: List[int], save_name: str):
+def plot_stacked_square_heated(datnums: List[int], save_name: str, plot=True):
     dats = get_dats(datnums)
 
     # Plot Integrated
@@ -264,6 +185,7 @@ def plot_stacked_square_heated(datnums: List[int], save_name: str):
         trace_name=lambda dat: f'Dat{dat.datnum}'
     )
 
+    figs = []
     for plot_info in [integrated_plot_info]:
         plotter = OneD(dats=dats)
         dat = dats[0]
@@ -294,7 +216,10 @@ def plot_stacked_square_heated(datnums: List[int], save_name: str):
                                         name=plot_info.trace_name(dat),
                                         hover_data=hover_data, hover_template=template,
                                         mode='lines'))
-        fig.show()
+        if plot:
+            fig.show()
+        figs.append(fig)
+    return figs
 
 
 def plot_vs_gamma(datnums: List[int], save_name: str, general: AnalysisGeneral, sf_name: Optional[str] = None,
@@ -343,6 +268,84 @@ def plot_transition_values(datnums: List[int], save_name: str, general: Analysis
     return fig
 
 
+def binned_y_fig(dat: DatHDF, save_name: str, num_chunks: int, which='entropy',
+                 integrated_info_name: Optional[str] = None,  # Only if not the same as save_name and want integrated
+                 ):
+    """Plotting data binned in y axis (after centering)"""
+
+    out = dat.SquareEntropy.get_Outputs(name=save_name, check_exists=True)
+    x = out.x
+    all_centers = out.centers_used
+    all_data = out.cycled  # Per row data
+    y = dat.Data.get_data('y')
+
+    binned = centered_y_bin(x, all_data, all_centers, num_bins=num_chunks,
+                            y=y)
+    binned_data = binned.binned_data
+    bin_centers = binned.y_centers
+
+    if which == 'entropy':
+        datas = [entropy_signal(d) for d in binned_data]
+        ylabel = f'{DELTA}I/nA'
+    elif which == 'transition':
+        datas = binned_data
+        ylabel = f'Current /nA'
+    elif which == 'integrated':
+        int_name = integrated_info_name if integrated_info_name else save_name
+        int_info = dat.Entropy.get_integration_info(int_name)
+        datas = [int_info.integrate(entropy_signal(d)) for d in binned_data]
+        ylabel = f'Entropy /kB'
+    else:
+        raise NotImplementedError
+
+    fig = get_slider_figure(datas=datas, xs=x,
+                            ids=[v for v in bin_centers],
+                            titles=[f'Dat{dat.datnum}: Bin centered at {v:.2f}' for v in bin_centers],
+                            xlabel=dat.Logs.xlabel, ylabel=ylabel)
+    return fig
+
+
+@dataclass
+class BinnedY:
+    """For storing return of center_y_bin"""
+    binned_data: List[np.ndarray]
+    avg_xs: List[np.ndarray]
+    y_centers: List[float]  # The center of the binned data in y-axis (or row num if y data not passed)
+
+
+def centered_y_bin(x: np.ndarray, data: np.ndarray,
+                   centers: Union[np.ndarray, Iterable], num_bins: int,
+                   y: Optional[np.ndarray] = None) -> BinnedY:
+    """
+    Bin 2D data in y axis after centering using centers.
+    Args:
+        x (): X array for all
+        data (): All data first
+        centers ():
+        num_bins ():
+        y ():
+
+    Returns:
+
+    """
+    if y is None:
+        y = range(data.shape[0])
+    bin_size = int(np.floor(data.shape[0] / num_bins))
+    binned_data = []
+    avg_xs = []
+    bin_centers = []
+    for i in range(int(data.shape[0] / bin_size)):
+        s = np.s_[i * bin_size:i * bin_size + bin_size]
+        d = data[s]
+        cs = centers[s]
+        bin_centers.append(float(np.nanmean(y[s])))
+        binned, x = U.mean_data(x, d, centers=cs, return_x=True)
+        binned_data.append(binned)
+        avg_xs.append(x)
+
+    return BinnedY(binned_data=binned_data, avg_xs=avg_xs, y_centers=bin_centers)
+
+
 # def plot_gamma_dcbias(datnums: List[int], save_name: str):
 #     if calculate:
 #         with ProcessPoolExecutor() as pool:
@@ -383,6 +386,10 @@ def make_long_analysis_params(entropy_datnums, transition_datnums, csq_datnums,
                               save_name: str,
                               entropy_data_rows: Tuple[Optional[int], Optional[int]] = (None, None),
                               transition_data_rows: Tuple[Optional[int], Optional[int]] = (None, None),
+                              transition_fit_width: Optional[float] = None,
+                              force_theta=None, force_gamma: Optional[int] = None,
+                              force_dt: Optional[float] = None, force_amp: Optional[float] = None,
+                              sf_from_square_transition: bool = True,
                               ) -> List[GammaAnalysisParams]:
     all_params = []
     for ed, td, cs in zip(entropy_datnums, transition_datnums, csq_datnums):
@@ -390,12 +397,12 @@ def make_long_analysis_params(entropy_datnums, transition_datnums, csq_datnums,
             GammaAnalysisParams(
                 save_name=save_name,
                 entropy_datnum=ed, transition_only_datnum=td, csq_datnum=cs,
-                setpoint_start=0.005,
+                setpoint_start=0.008,
                 entropy_transition_func_name='i_sense', entropy_fit_width=None, entropy_data_rows=entropy_data_rows,
-                force_dt=None, force_amp=None,
-                sf_from_square_transition=True,
-                force_theta=None, force_gamma=None,  # Applies to entropy and transition only
-                transition_func_name='i_sense_digamma', transition_fit_width=None,
+                force_dt=force_dt, force_amp=force_amp,
+                sf_from_square_transition=sf_from_square_transition,
+                force_theta=force_theta, force_gamma=force_gamma,  # Applies to entropy and transition only
+                transition_func_name='i_sense_digamma', transition_fit_width=transition_fit_width,
                 transition_center_func_name='i_sense_digamma', transition_data_rows=transition_data_rows,
                 csq_mapped=True
             )
@@ -456,9 +463,15 @@ if __name__ == '__main__':
     name = 'forced_theta_linear'
     # sf_name = 'fixed_dT'
     sf_name = 'scaled_dT'
-    all_params = make_vs_gamma_analysis_params(VS_GAMMA, VS_GAMMA_Tonly, save_name=name,
-                                               force_theta=None, force_gamma=None,
-                                               sf_from_square_transition=False, width=600)
+    # all_params = make_vs_gamma_analysis_params(VS_GAMMA, VS_GAMMA_Tonly, save_name=name,
+    #                                            force_theta=3.9, force_gamma=None,
+    #                                            sf_from_square_transition=False, width=600)
+    all_params = make_long_analysis_params(LONG_GAMMA, LONG_GAMMA_Tonly, LONG_GAMMA_csq, save_name=name,
+                                           force_theta=None, force_gamma=None,  # Theta set below
+                                           transition_fit_width=500,
+                                           force_dt=1.11,
+                                           sf_from_square_transition=False,
+                                           )
 
     # Setting theta according to linear fit in weakly coupled regime with gamma = 0
     line = lm.models.LinearModel()
@@ -475,7 +488,8 @@ if __name__ == '__main__':
             par.force_amp = 0.52
 
     # Do processing
-    general = AnalysisGeneral(params_list=all_params, calculate=False, overwrite_entropy=True, overwrite_transition=True)
+    general = AnalysisGeneral(params_list=all_params, calculate=True, overwrite_entropy=False,
+                              overwrite_transition=False)
     run_processing(general)
     for par in all_params:
         calculate_new_sf_only(entropy_datnum=par.entropy_datnum, save_name=sf_name,
@@ -483,27 +497,65 @@ if __name__ == '__main__':
                               from_square_transition=par.sf_from_square_transition,
                               transition_datnum=par.transition_only_datnum,
                               fit_name=name)
+    #
+    # name = 'temp'
+    # params = make_long_analysis_params([2164], transition_datnums=[2165],csq_datnums=[2166], save_name=name,
+    #                                        force_theta=3.9, force_gamma=None,
+    #                                        sf_from_square_transition=False,
+    #                                        transition_fit_width=500)[0]
+    # params.force_dt = 1.11
+    #
+    # setup_csq_dat(2166)
+    # process_single(params, overwrite_transition=False, overwrite_entropy=False)
+
+    dat = get_dat(2213)
+
+    out = dat.SquareEntropy.get_Outputs(name=name, check_exists=True)
+    x = out.x
+    all_centers = out.centers_used
+    all_data = out.cycled  # Per row data
+    y = dat.Data.get_data('y')
+
+    binned = centered_y_bin(x, all_data, all_centers, num_bins=10,
+                            y=y)
+
+    int_info = dat.Entropy.get_integration_info(name=name)
+    plotter = OneD(dat=dat)
+    integrated_entropies = [int_info.integrate(entropy_signal(d))[-1] for d in binned.binned_data]
+    x = binned.y_centers
+    fig = plotter.plot(data=integrated_entropies, x=x, xlabel=dat.Logs.ylabel, ylabel='Entropy /kB', title=f'Dat{dat.datnum}: Integrated entropy for binned data')
+    fig.show()
+    # fits = dat.SquareEntropy.get_row_fits(name='temp')
+
+    # for n in ['entropy', 'transition', 'integrated']:
+    for n in ['integrated']:
+        fig = binned_y_fig(dat, save_name=name, num_chunks=10, which=n, integrated_info_name=name)
+        fig.show()
+        if n == 'integrated':
+            fig.write_html(f'figs/Dat{dat.datnum}-slider_integrated.html')
 
     # Plotting
-    # plot_stacked_square_heated(LONG_GAMMA, save_name='test')
+    figs = plot_stacked_square_heated(LONG_GAMMA, save_name=name)
+    for i, fig in enumerate(figs):
+        fig.write_html(f'temp{i}.html')
 
     # fig = get_integrated_fig(get_dats(general.entropy_datnums), title_append=f'comparing scaling factors')
-    # for int_name in ['scaled_dT', 'fixed_dT']:
+    # # for int_name in ['scaled_dT', 'fixed_dT']:
+    # for int_name in ['scaled_dT']:
     #     fig.add_trace(get_integrated_trace(dats=get_dats(general.entropy_datnums),
-    #                                    x_func=general.x_func, x_label=general.x_label,
-    #                                    trace_name=int_name,
-    #                                    fit_name=name,
-    #                                    int_info_name=int_name, SE_output_name=name))
+    #                                x_func=general.x_func, x_label=general.x_label,
+    #                                trace_name=int_name,
+    #                                fit_name=name,
+    #                                int_info_name=int_name, SE_output_name=name))
     # fig.show()
-
-
+    #
     # fig = plot_transition_values(general.transition_datnums, save_name=name, general=general, param_name='theta',
     #                        transition_only=True)
     # fig = plot_transition_values(general.transition_datnums, save_name=name, general=general, param_name='g',
     #                        transition_only=True)
     # fig = plot_transition_values(general.transition_datnums, save_name=name, general=general, param_name='amp',
     #                        transition_only=True)
-
+    #
     # fig = plot_transition_values(general.transition_datnums, save_name=name, general=general, param_name='theta',
     #                              transition_only=True, show=False)
     # line = lm.models.LinearModel()
@@ -516,3 +568,16 @@ if __name__ == '__main__':
     # fig.add_trace(plotter.trace(fit.eval(x=(x := np.linspace(-380, -180, 101))), x=x, mode='lines', name='Fit'))
     # print(fit.fit_report())
     # fig.show()
+
+# import h5py
+# def test(f: h5py.File, shape=(1000,), dtype='S10'):
+#     if 'test' in f.keys():
+#         del f['test']
+#     f.create_dataset('test', shape, dtype=dtype)
+#     store = [b'blablablablabla']*70
+#     f['test'][:70] = store
+#     read = f['test'][:]
+#     return read
+#     # del f['test']
+#
+
