+++ new_dash/pages/single_entropy.py
@@ -1,35 +1,31 @@
 from __future__ import annotations
-from typing import List, Tuple, Dict, Optional, Any, Callable, Union
+from typing import List, Tuple, Dict, Optional, Callable, Union
 import abc
 from dataclasses import dataclass
 from functools import partial
 import logging
 
-from dash_dashboard.base_classes import BasePageLayout, BaseMain, BaseSideBar, PageInteractiveComponents, \
+from dash_dashboard.base_classes import PageInteractiveComponents, \
     CommonInputCallbacks, PendingCallbacks
 from dash_dashboard.util import triggered_by
 from new_dash.base_class_overrides import DatDashPageLayout, DatDashMain, DatDashSidebar
 import dash_dashboard.component_defaults as c
-from dash_extensions.enrich import ServersideOutput
 
 import dash_html_components as html
 import dash_bootstrap_components as dbc
-from dash_extensions.enrich import MultiplexerTransform  # Dash Extensions has some super useful things!
 from dash import no_update
 from dash.exceptions import PreventUpdate
 import plotly.graph_objects as go
 
-from src.DatObject.Make_Dat import get_dat, get_dats, DatHDF
+from src.DatObject.Make_Dat import get_dat, DatHDF
 from src.Dash.DatPlotting import OneD, TwoD
 import src.UsefulFunctions as U
-from src.AnalysisTools.fitting import FitInfo, calculate_fit
+from src.AnalysisTools.fitting import CalculatedTransitionFit, CalculatedEntropyFit, \
+    calculate_se_output, calculate_tonly_data, TransitionCalcParams, set_centers, _get_data_in_range, \
+    _get_transition_fit_func_params
 
-from Analysis.Feb2021.entropy_gamma_final import GammaAnalysisParams
-from src.DatObject.Attributes.SquareEntropy import Output, centers_from_fits
-from Analysis.Feb2021.common import _get_transition_fit_func_params, square_wave_time_array, _get_data_in_range, \
-    calculate_csq_map
-from src.DatObject.Attributes.SquareEntropy import Input as SeInput, Output as SeOutput, ProcessParams as SePars, \
-    process_per_row_parts
+from src.AnalysisTools.gamma_entropy import GammaAnalysisParams
+from src.DatObject.Attributes.SquareEntropy import Output
 from src.DatObject.Attributes.Entropy import IntegrationInfo, scaling
 
 import numpy as np
@@ -104,6 +100,11 @@ class Components(PageInteractiveComponents):
         self.graph_2 = c.graph_area(id_name='graph-2', graph_header='Transition',
                                     pending_callbacks=self.pending_callbacks)
         self.graph_3 = c.graph_area(id_name='graph-3', graph_header='Integrated',
+                                    pending_callbacks=self.pending_callbacks)
+
+        self.graph_4 = c.graph_area(id_name='graph-4', graph_header='2D Entropy',
+                                    pending_callbacks=self.pending_callbacks)
+        self.graph_5 = c.graph_area(id_name='graph-5', graph_header='2D Transition',
                                     pending_callbacks=self.pending_callbacks)
 
         # Info Area
@@ -240,6 +241,8 @@ class SingleEntropyMain(DatDashMain, abc.ABC):
                     html.H6('Integrated Info'),
                     self.components.table_3,
                     html.Hr(),
+                    self.components.graph_4,
+                    self.components.graph_5,
 
                 ], width=4)
             ])
@@ -251,7 +254,10 @@ class SingleEntropyMain(DatDashMain, abc.ABC):
         # Graph Callbacks
         for graph, cb_func in {components.graph_1: 'entropy_signal',
                                components.graph_2: 'transition_data',
-                               components.graph_3: 'integrated_entropy'}.items():
+                               components.graph_3: 'integrated_entropy',
+                               components.graph_4: '2d_entropy',
+                               components.graph_5: '2d_transition',
+                               }.items():
             self.make_callback(outputs=(graph.graph_id, 'figure'),
                                inputs=GraphCallbacks.get_inputs(),
                                func=GraphCallbacks.get_callback_func(cb_func),
@@ -482,6 +488,9 @@ class GraphCallbacks(CommonInputCallbacks):
             "entropy_signal": self.entropy_signal,
             "transition_data": self.transition_data,
             "integrated_entropy": self.integrated_entropy,
+            "2d_entropy": self.entropy_2d,
+            "2d_transition": self.transition_2d,
+
         }
 
     def _correct_call_args(self) -> bool:
@@ -489,6 +498,25 @@ class GraphCallbacks(CommonInputCallbacks):
         if any([self.dat is None, not is_square_entropy_dat(self.dat)]):
             return False
         return True
+
+    def entropy_2d(self) -> go.Figure:
+        if not self._correct_call_args():
+            logger.warning(f'Bad call args to GraphCallback')
+            return go.Figure()
+        dat = self.dat
+        plotter = TwoD(dat=dat)
+        if self.calculated_triggered:
+            pars = self.calculated.analysis_params
+            if pars.
+            fig = plotter.plot()
+
+
+
+
+        return go.Figure()
+
+    def transition_2d(self) -> go.Figure:
+        return go.Figure()
 
     def entropy_signal(self) -> go.Figure:
         """dN/dT figure"""
@@ -500,8 +528,12 @@ class GraphCallbacks(CommonInputCallbacks):
         plotter = OneD(dat=dat)
         fig = plotter.figure(title=f'Dat{dat.datnum}: dN/dT')
         out = dat.SquareEntropy.get_Outputs(name=self.se_name, check_exists=True)
-        x = out.x
-        data = out.average_entropy_signal
+        if self.calculated_triggered:
+            x = self.calculated.calculated_entropy_fit.x
+            data = self.calculated.calculated_entropy_fit.data
+        else:
+            x = out.x
+            data = out.average_entropy_signal
         fig.add_trace(plotter.trace(data=data, x=x, mode='lines', name='Data'))
         existing_names = dat.Entropy.fit_names
         for n in self.e_fit_names:
@@ -524,9 +556,13 @@ class GraphCallbacks(CommonInputCallbacks):
         dat = self.dat
         plotter = OneD(dat=dat)
         fig = plotter.figure(title=f'Dat{dat.datnum}: Transition')
-        out = dat.SquareEntropy.get_Outputs(name=self.se_name, check_exists=True)
-        x = out.x
-        datas = out.averaged
+        if self.calculated_triggered:
+            x = self.calculated.calculated_transition_fit.x
+            datas = self.calculated.calculated_transition_fit.data
+        else:
+            out = dat.SquareEntropy.get_Outputs(name=self.se_name, check_exists=True)
+            x = out.x
+            datas = out.averaged
         biases = dat.SquareEntropy.square_awg.AWs[0][0]
         for data, label in zip(datas, ['0nA_0', f'{biases[1] / 10:.1f}nA', '0nA_1', f'{biases[3] / 10:.1f}nA']):
             fig.add_trace(plotter.trace(data=data, x=x, mode='lines', name=label))
@@ -550,9 +586,13 @@ class GraphCallbacks(CommonInputCallbacks):
         dat = self.dat
         plotter = OneD(dat=dat)
         fig = plotter.figure(title=f'Dat{dat.datnum}: Integrated')
-        out = dat.SquareEntropy.get_Outputs(name=self.se_name, check_exists=True)
-        x = out.x
-        data = out.average_entropy_signal
+        if self.calculated_triggered:
+            x = self.calculated.calculated_entropy_fit.x
+            data = self.calculated.calculated_entropy_fit.data
+        else:
+            out = dat.SquareEntropy.get_Outputs(name=self.se_name, check_exists=True)
+            x = out.x
+            data = out.average_entropy_signal
         existing_names = dat.Entropy.get_integration_info_names()
         for n in self.int_names:
             if n in existing_names:
@@ -978,23 +1018,6 @@ class CalculateCallback(CommonInputCallbacks):
 
 
 @dataclass
-class CalculatedFit:
-    x: np.ndarray
-    data: np.ndarray
-    fit: FitInfo
-
-
-@dataclass
-class CalculatedTransitionFit(CalculatedFit):
-    pass
-
-
-@dataclass
-class CalculatedEntropyFit(CalculatedFit):
-    pass
-
-
-@dataclass
 class StoreData:
     analysis_params: GammaAnalysisParams
     SE_output: Output
@@ -1042,185 +1065,6 @@ def callbacks(app):
 
 
 # ###########################################
-
-def calculate_se_output(dat: DatHDF, rows, csq_mapped,
-                        center_func_name: str,
-                        setpoint_start: Optional[float]) -> SeOutput:
-    """
-    Calculate SE output using EXISTING center fits... I.e. no fitting run in this, just the SE processing and averaging
-    using existing center fits.
-
-    Args:
-        dat (): SE entropy dat
-        rows (): Rows to process between (Nones are beginning or end)
-        csq_mapped (): Whether to use CSQ mapped data
-        center_func_name (): Name of func used for centering (these fits will be loaded from)
-        setpoint_start (): Amount of time after each setpoint to throw out
-
-    Returns:
-        SeOutput with all relevant data filled
-    """
-
-    def get_setpoint_ids(d: DatHDF, start_time, fin_time=None):
-        sps = [start_time, fin_time]
-        sp_times = square_wave_time_array(d.SquareEntropy.square_awg)
-        start, fin = [U.get_data_index(sp_times, sp) for sp in sps]
-        return start, fin
-
-    def get_data(d: DatHDF, rs, csq):
-        s, f = rs
-        if csq:
-            data_ = d.Data.get_data('csq_mapped')[s:f]
-        else:
-            data_ = d.Transition.get_data('i_sense')[s:f]
-        return data_
-
-    centers = get_centers(dat, center_func_name=center_func_name, rows=rows, se_data=True)
-
-    sp_start, sp_fin = get_setpoint_ids(dat, setpoint_start, None)
-    x = dat.Data.get_data('x')
-    data = get_data(dat, rows, csq_mapped)
-    inputs = dat.SquareEntropy.get_Inputs(name=None, x_array=x, i_sense=data, centers=centers,
-                                          save_name=None)
-
-    process_params = dat.SquareEntropy.get_ProcessParams(name=None,
-                                                         setpoint_start=sp_start, setpoint_fin=sp_fin,
-                                                         transition_fit_func=None,  # Don't need to center again
-                                                         transition_fit_params=None,  # Don't need to center again
-                                                         save_name=None,  # Do not save
-                                                         )
-    out = dat.SquareEntropy.get_Outputs(inputs=inputs, process_params=process_params,
-                                        calculate_only=True)
-    return out
-
-
-def calculate_tonly_data(dat: DatHDF, rows, csq_mapped,
-                         center_func_name: str) -> Tuple[np.ndarray, np.ndarray]:
-    """
-    Calculate Averaged Transition data for rows selected using named center func
-
-    Args:
-        dat ():
-        rows ():
-        csq_mapped ():
-        center_func_name ():
-
-    Returns:
-
-    """
-
-    def get_data(rs, csq, transition_dat: DatHDF) -> Tuple[np.ndarray, np.ndarray]:
-        if csq:
-            name = 'csq_mapped'
-            data_group_name = 'Data'
-        else:
-            name = 'i_sense'
-            data_group_name = 'Transition'
-        s, f = rs
-        x_ = transition_dat.Data.get_data('x', data_group_name=data_group_name)
-        data_ = transition_dat.Data.get_data(name, data_group_name=data_group_name)[s:f]
-        return x_, data_
-
-    x, data = get_data(rows, csq_mapped, dat)
-    centers = get_centers(dat, center_func_name=center_func_name, rows=rows, se_data=False)
-
-    data_avg, x_avg = U.mean_data(x=x, data=data, centers=centers, method='linear', return_x=True)
-    return x_avg, data_avg
-
-
-@dataclass
-class TransitionCalcParams:
-    initial_x: np.ndarray  # For getting param estimates
-    initial_data: np.ndarray  # For getting param estimates (1D)
-    force_theta: Optional[float]
-    force_gamma: Optional[float]
-    csq_mapped: bool = False
-
-
-def get_centers(dat: DatHDF, center_func_name: str, rows: Tuple[Optional[float], Optional[float]],
-                se_data: bool = False) -> np.ndarray:
-    """
-
-    Args:
-        dat (): For getting data from (and potentially saving new fits to if check_exists = False)
-        center_func_name (): which transition func as string
-        rows (): For rows between
-        se_data (): Use SE fits instead of Transition fits (i.e. dat.SquareEntropy vs dat.Transition)
-
-    Returns:
-        array of centers
-    """
-
-    def get_fit_name(f_name: str) -> str:
-        return 'centering_' + f_name
-
-    fit_name = get_fit_name(center_func_name)
-    rows = (rows[0] if rows[0] else 0, rows[1] if rows[1] else dat.Data.get_data('y').shape[0])
-
-    if se_data:
-        row_fits = [dat.SquareEntropy.get_fit(which_fit='transition', which='row', row=r,
-                                              fit_name=fit_name) for r in range(*rows)]
-    else:
-        row_fits = [dat.Transition.get_fit(which='row', row=r, name=fit_name) for r in range(*rows)]
-
-    return centers_from_fits(row_fits)
-
-
-def set_centers(dat: DatHDF, center_func_name: str, calc_params: Optional[TransitionCalcParams] = None,
-                se_data: bool = False, csq_mapped: bool = False) -> np.ndarray:
-    """
-
-    Args:
-        dat (): For getting data from (and potentially saving new fits to if check_exists = False)
-        center_func_name (): which transition func as string
-        se_data (): Use SE fits instead of Transition fits (i.e. dat.SquareEntropy vs dat.Transition)
-        csq_mapped: Whether to fit regular or csq_mapped data
-        calc_params (): Used only if check_exists = False
-
-    Returns:
-        array of centers
-    """
-
-    def get_fit_name(f_name: str) -> str:
-        return 'centering_' + f_name
-
-    def get_data(d: DatHDF, csq):
-        if csq:
-            data_ = d.Data.get_data('csq_mapped')
-        else:
-            data_ = d.Data.get_data('i_sense')
-        return data_
-
-    fit_name = get_fit_name(center_func_name)
-
-    cp = calc_params
-    fit_func, params = _get_transition_fit_func_params(x=cp.initial_x, data=cp.initial_data,
-                                                       t_func_name=center_func_name,
-                                                       theta=cp.force_theta, gamma=cp.force_gamma)
-
-    x = dat.Data.get_data('x')
-    data = get_data(dat, csq_mapped)
-    if se_data:
-        name = 'csq_mapped cycled only' if csq_mapped else 'i_sense cycled only'
-        inputs = dat.SquareEntropy.get_Inputs(x_array=x, i_sense=data)
-        # Calculate row only output if necessary (overwrite = False)
-        pre_out = dat.SquareEntropy.get_row_only_output(name=name, inputs=inputs, process_params=None,
-                                                        check_exists=False, overwrite=False)
-        row_fits = [dat.SquareEntropy.get_fit(which_fit='transition', which='row', row=i,
-                                              x=pre_out.x,
-                                              data=d,
-                                              fit_name=fit_name,
-                                              check_exists=False,
-                                              initial_params=params, fit_func=fit_func,
-                                              transition_part='cold') for i, d in enumerate(pre_out.cycled)]
-    else:
-        row_fits = [dat.Transition.get_fit(which='row', row=i, name=fit_name,
-                                           check_exists=False,
-                                           x=x,
-                                           data=d,
-                                           initial_params=params, fit_func=fit_func) for i, d in enumerate(data)]
-
-    return centers_from_fits(row_fits)
 
 
 if __name__ == '__main__':
