+++ src/AnalysisTools/fitting.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 import inspect
 from dataclasses import dataclass, InitVar, field
 from hashlib import md5
-from typing import Union, Optional, Callable, Any, TYPE_CHECKING
+from typing import Union, Optional, Callable, Any, TYPE_CHECKING, Tuple
 
 import h5py
 import lmfit as lm
@@ -10,8 +10,16 @@ import numpy as np
 import pandas as pd
 import logging
 
-from src import CoreUtil as CU
+import scipy
+from scipy.interpolate import RectBivariateSpline
+
+from src import CoreUtil as CU, UsefulFunctions as U
+from src.DatObject.Attributes.SquareEntropy import Output as SeOutput, square_wave_time_array, centers_from_fits
+from src.DatObject.Attributes.Transition import get_param_estimates, get_transition_function
 from src.HDF_Util import params_from_HDF, params_to_HDF, NotFoundInHdfError, DatDataclassTemplate
+
+if TYPE_CHECKING:
+    from src.DatObject.DatHDF import DatHDF
 
 logger = logging.getLogger(__name__)
 
@@ -342,3 +350,320 @@ def calculate_fit(x: np.ndarray, data: np.ndarray, params: lm.Parameters, func: 
         fit = None
     return fit
 
+
+@dataclass
+class CalculatedFit:
+    x: np.ndarray
+    data: np.ndarray
+    fit: FitInfo
+
+
+@dataclass
+class CalculatedTransitionFit(CalculatedFit):
+    pass
+
+
+@dataclass
+class CalculatedEntropyFit(CalculatedFit):
+    pass
+
+
+def calculate_se_output(dat: DatHDF, rows, csq_mapped,
+                        center_func_name: str,
+                        setpoint_start: Optional[float]) -> SeOutput:
+    """
+    Calculate SE output using EXISTING center fits... I.e. no fitting run in this, just the SE processing and averaging
+    using existing center fits.
+
+    Args:
+        dat (): SE entropy dat
+        rows (): Rows to process between (Nones are beginning or end)
+        csq_mapped (): Whether to use CSQ mapped data
+        center_func_name (): Name of func used for centering (these fits will be loaded from)
+        setpoint_start (): Amount of time after each setpoint to throw out
+
+    Returns:
+        SeOutput with all relevant data filled
+    """
+
+    def get_setpoint_ids(d: DatHDF, start_time, fin_time=None):
+        sps = [start_time, fin_time]
+        sp_times = square_wave_time_array(d.SquareEntropy.square_awg)
+        start, fin = [U.get_data_index(sp_times, sp) for sp in sps]
+        return start, fin
+
+    def get_data(d: DatHDF, rs, csq):
+        s, f = rs
+        if csq:
+            data_ = d.Data.get_data('csq_mapped')[s:f]
+        else:
+            data_ = d.Transition.get_data('i_sense')[s:f]
+        return data_
+
+    centers = get_centers(dat, center_func_name=center_func_name, rows=rows, se_data=True)
+
+    sp_start, sp_fin = get_setpoint_ids(dat, setpoint_start, None)
+    x = dat.Data.get_data('x')
+    data = get_data(dat, rows, csq_mapped)
+    inputs = dat.SquareEntropy.get_Inputs(name=None, x_array=x, i_sense=data, centers=centers,
+                                          save_name=None)
+
+    process_params = dat.SquareEntropy.get_ProcessParams(name=None,
+                                                         setpoint_start=sp_start, setpoint_fin=sp_fin,
+                                                         transition_fit_func=None,  # Don't need to center again
+                                                         transition_fit_params=None,  # Don't need to center again
+                                                         save_name=None,  # Do not save
+                                                         )
+    out = dat.SquareEntropy.get_Outputs(inputs=inputs, process_params=process_params,
+                                        calculate_only=True)
+    return out
+
+
+def calculate_tonly_data(dat: DatHDF, rows, csq_mapped,
+                         center_func_name: str) -> Tuple[np.ndarray, np.ndarray]:
+    """
+    Calculate Averaged Transition data for rows selected using named center func
+
+    Args:
+        dat ():
+        rows ():
+        csq_mapped ():
+        center_func_name ():
+
+    Returns:
+
+    """
+
+    def get_data(rs, csq, transition_dat: DatHDF) -> Tuple[np.ndarray, np.ndarray]:
+        if csq:
+            name = 'csq_mapped'
+            data_group_name = 'Data'
+        else:
+            name = 'i_sense'
+            data_group_name = 'Transition'
+        s, f = rs
+        x_ = transition_dat.Data.get_data('x', data_group_name=data_group_name)
+        data_ = transition_dat.Data.get_data(name, data_group_name=data_group_name)[s:f]
+        return x_, data_
+
+    x, data = get_data(rows, csq_mapped, dat)
+    centers = get_centers(dat, center_func_name=center_func_name, rows=rows, se_data=False)
+
+    data_avg, x_avg = U.mean_data(x=x, data=data, centers=centers, method='linear', return_x=True)
+    return x_avg, data_avg
+
+
+@dataclass
+class TransitionCalcParams:
+    initial_x: np.ndarray  # For getting param estimates
+    initial_data: np.ndarray  # For getting param estimates (1D)
+    force_theta: Optional[float]
+    force_gamma: Optional[float]
+    csq_mapped: bool = False
+
+
+def get_centers(dat: DatHDF, center_func_name: str, rows: Tuple[Optional[float], Optional[float]],
+                se_data: bool = False) -> np.ndarray:
+    """
+
+    Args:
+        dat (): For getting data from (and potentially saving new fits to if check_exists = False)
+        center_func_name (): which transition func as string
+        rows (): For rows between
+        se_data (): Use SE fits instead of Transition fits (i.e. dat.SquareEntropy vs dat.Transition)
+
+    Returns:
+        array of centers
+    """
+
+    def get_fit_name(f_name: str) -> str:
+        return 'centering_' + f_name
+
+    fit_name = get_fit_name(center_func_name)
+    rows = (rows[0] if rows[0] else 0, rows[1] if rows[1] else dat.Data.get_data('y').shape[0])
+
+    if se_data:
+        row_fits = [dat.SquareEntropy.get_fit(which_fit='transition', which='row', row=r,
+                                              fit_name=fit_name) for r in range(*rows)]
+    else:
+        row_fits = [dat.Transition.get_fit(which='row', row=r, name=fit_name) for r in range(*rows)]
+
+    return centers_from_fits(row_fits)
+
+
+def set_centers(dat: DatHDF, center_func_name: str, calc_params: Optional[TransitionCalcParams] = None,
+                se_data: bool = False, csq_mapped: bool = False) -> np.ndarray:
+    """
+
+    Args:
+        dat (): For getting data from (and potentially saving new fits to if check_exists = False)
+        center_func_name (): which transition func as string
+        se_data (): Use SE fits instead of Transition fits (i.e. dat.SquareEntropy vs dat.Transition)
+        csq_mapped: Whether to fit regular or csq_mapped data
+        calc_params (): Used only if check_exists = False
+
+    Returns:
+        array of centers
+    """
+
+    def get_fit_name(f_name: str) -> str:
+        return 'centering_' + f_name
+
+    def get_data(d: DatHDF, csq):
+        if csq:
+            data_ = d.Data.get_data('csq_mapped')
+        else:
+            data_ = d.Data.get_data('i_sense')
+        return data_
+
+    fit_name = get_fit_name(center_func_name)
+
+    cp = calc_params
+    fit_func, params = _get_transition_fit_func_params(x=cp.initial_x, data=cp.initial_data,
+                                                       t_func_name=center_func_name,
+                                                       theta=cp.force_theta, gamma=cp.force_gamma)
+
+    x = dat.Data.get_data('x')
+    data = get_data(dat, csq_mapped)
+    if se_data:
+        name = 'csq_mapped cycled only' if csq_mapped else 'i_sense cycled only'
+        inputs = dat.SquareEntropy.get_Inputs(x_array=x, i_sense=data)
+        # Calculate row only output if necessary (overwrite = False)
+        pre_out = dat.SquareEntropy.get_row_only_output(name=name, inputs=inputs, process_params=None,
+                                                        check_exists=False, overwrite=False)
+        row_fits = [dat.SquareEntropy.get_fit(which_fit='transition', which='row', row=i,
+                                              x=pre_out.x,
+                                              data=d,
+                                              fit_name=fit_name,
+                                              check_exists=False,
+                                              initial_params=params, fit_func=fit_func,
+                                              transition_part='cold') for i, d in enumerate(pre_out.cycled)]
+    else:
+        row_fits = [dat.Transition.get_fit(which='row', row=i, name=fit_name,
+                                           check_exists=False,
+                                           x=x,
+                                           data=d,
+                                           initial_params=params, fit_func=fit_func) for i, d in enumerate(data)]
+
+    return centers_from_fits(row_fits)
+
+
+def get_default_transition_params(func_name: str,
+                                  x: Optional[np.ndarray] = None, data: Optional[np.ndarray] = None) -> lm.Parameters:
+    params = get_param_estimates(x=x, data=data)
+    if func_name == 'i_sense_digamma':
+        params.add('g', 0, min=-50, max=1000, vary=True)
+    elif func_name == 'i_sense_digamma_amplin':
+        params.add('g', 0, min=-50, max=1000, vary=True)
+        params.add('amplin', 0, vary=True)
+    return params
+
+
+def calculate_transition_only_fit(datnum, save_name, t_func_name: str = 'i_sense_digamma', theta=None, gamma=None,
+                                  x: Optional[np.ndarray] = None, data: Optional[np.ndarray] = None,
+                                  width: Optional[float] = None, center: Optional[float] = None,
+                                  overwrite=False) -> FitInfo:
+    from src.DatObject.Make_Dat import get_dat
+    dat = get_dat(datnum)
+
+    x = x if x is not None else dat.Transition.avg_x
+    data = data if data is not None else dat.Transition.avg_data
+
+    x, data = _get_data_in_range(x, data, width, center=center)
+
+    t_func, params = _get_transition_fit_func_params(x, data, t_func_name, theta, gamma)
+
+    return dat.Transition.get_fit(name=save_name, fit_func=t_func,
+                                  data=data, x=x, initial_params=params,
+                                  check_exists=False, overwrite=overwrite)
+
+
+def _get_data_in_range(x: np.ndarray, data: np.ndarray, width: Optional[float], center: Optional[float] = None) -> \
+        Tuple[np.ndarray, np.ndarray]:
+    if center is None:
+        center = 0
+    if width is not None:
+        x, data = np.copy(x), np.copy(data)
+
+        start_ind = np.nanargmin(np.abs(np.add(x, width + center)))
+        end_ind = np.nanargmin(np.abs(np.subtract(x, width + center)))
+
+        x[:start_ind] = [np.nan] * start_ind
+        x[end_ind:] = [np.nan] * (len(x) - end_ind)
+
+        data[:start_ind] = [np.nan] * start_ind
+        data[end_ind:] = [np.nan] * (len(data) - end_ind)
+    return x, data
+
+
+def _get_transition_fit_func_params(x, data, t_func_name, theta, gamma):
+    """
+
+    Args:
+        x ():
+        data ():
+        t_func_name ():
+        theta ():
+        gamma ():
+
+    Returns:
+
+    """
+    t_func = get_transition_function(t_func_name)
+    params = get_default_transition_params(t_func_name, x, data)
+    if theta:
+        params = U.edit_params(params, 'theta', value=theta, vary=False)
+    if gamma is not None and 'g' in params:
+        params = U.edit_params(params, 'g', gamma, False)
+    return t_func, params
+
+
+def calculate_se_transition(datnum: int, save_name: str, se_output_name: str, t_func_name: str = 'i_sense_digamma',
+                            theta=None, gamma=None,
+                            transition_part: str = 'cold',
+                            width: Optional[float] = None, center: Optional[float] = None,
+                            overwrite=False):
+    from src.DatObject.Make_Dat import get_dat
+    dat = get_dat(datnum)
+    data = dat.SquareEntropy.get_transition_part(name=se_output_name, part=transition_part, existing_only=True)
+    x = dat.SquareEntropy.get_Outputs(name=se_output_name, check_exists=True).x
+
+    x, data = _get_data_in_range(x, data, width, center=center)
+
+    t_func, params = _get_transition_fit_func_params(x, data, t_func_name, theta, gamma)
+
+    return dat.SquareEntropy.get_fit(which_fit='transition', transition_part=transition_part, fit_name=save_name,
+                                     fit_func=t_func, initial_params=params, data=data, x=x, check_exists=False,
+                                     overwrite=overwrite)
+
+
+def calculate_se_entropy_fit(datnum: int, save_name: str, se_output_name: str,
+                             width: Optional[float] = None, center: Optional[float] = None,
+                             overwrite=False):
+    from src.DatObject.Make_Dat import get_dat
+    dat = get_dat(datnum)
+    out = dat.SquareEntropy.get_Outputs(name=se_output_name, check_exists=True)
+    x = out.x
+    data = out.average_entropy_signal
+
+    x, data = _get_data_in_range(x, data, width, center)
+    return dat.Entropy.get_fit(name=save_name, x=out.x, data=data, check_exists=False, overwrite=overwrite)
+
+
+def NRG_fitter() -> Callable:
+    NRG = scipy.io.loadmat('NRGResults.mat')
+    occ = NRG["Occupation_mat"]
+    ens = np.reshape(NRG["Ens"], 401)
+    ts = np.reshape(NRG["Ts"], 70)
+    ens = np.flip(ens)
+    occ = np.flip(occ, 0)
+    interp = RectBivariateSpline(ens, np.log10(ts), occ, kx=1, ky=1)
+
+    def interpNRG(x, logt, dx=1, amp=1, center=0, lin=0, const=0):
+        ens = np.multiply(np.add(x, center), dx)
+        curr = [interp(en, logt)[0][0] for en in ens]
+        scaled_current = np.multiply(curr, amp)
+        scaled_current += const + np.multiply(lin, x)
+        return scaled_current
+
+    return interpNRG
